backend/pyproject.toml
[project]
name = "estate-scout-backend"
version = "0.1.0"
description = "Estate Scout API - estate sale aggregator backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "sqlalchemy[asyncio]==2.0.30",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
    "pydantic-settings==2.2.1",
    "redis[hiredis]==5.0.4",
    "httpx>=0.28.0",
    "brotli>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "python-dateutil>=2.9.0",
    "pgvector==0.2.4",
    "openai==1.30.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.9",
    "aiofiles==23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
